==> scripts/30_fitModel.R <==
source('../scripts/functions.R')


## Fit the initial model that will serve as a broad variable selection
## process
source('../scripts/31_fitModel_gbm.R')
## Fit temporal models
source('../scripts/32_fitModel_gbm_temporal.R')
